Refactor DatabaseConnector and ProxySQLExecutor for multi logic sql transaction control
